refactor(ingest): hoist normalize_claim_path to common::paths (refs #43)
